improve code

+++ Dashboard.Application/Features/Projects/DeleteProject/DeleteProjectCommand.cs
@@ -1,5 +1,6 @@
 using Dashboard.BuildingBlock.Exceptions;
 using Dashboard.BuildingBlock.Repository;
+using Dashboard.Domain.Enums;
 using Dashboard.Domain.ProjectDomain;
 using MediatR;
 
@@ -11,6 +12,11 @@ public class DeleteProjectCommand(IRepository<Project> repository) : IRequestHan
     {
         var project = await repository.GetByIdAsync(request.Id, cancellationToken)
                       ?? throw new EntityNotFoundException("Project not found");
+
+        if (project.Status == ProjectStatus.Completed || project.Status == ProjectStatus.Cancelled)
+        {
+            throw new InvalidProjectException("Can't delete completed project");
+        }
 
         var result = await repository.DeleteAsync(project, cancellationToken);
         return result;
